data is posted correctly just a syntax error missing

+++ server/db/migrations/20240330031735_portfolios.js
@@ -8,11 +8,11 @@ export async function up(knex) {
     table.integer('authO_id')
     table.string('coin_id')
     table.string('coin_name')
-    table.decimal('price')
-    table.decimal('percent_change_1h')
-    table.decimal('percent_change_24h')
-    table.decimal('percent_change_7d')
-    table.decimal('market_cap')
+    table.integer('price')
+    table.integer('percent_change_1h')
+    table.integer('percent_change_24h')
+    table.integer('percent_change_7d')
+    table.integer('market_cap')
   })
 }
